requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.23.3
supabase==2.3.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv==1.0.0
python-multipart==0.0.6
